prevent exceptions from throwing.
I have an empty disc drive and so it would throw every time I run this code.

change to std::error_code version of std::filesystem::exists

to silently handle the error.

was missing a semicolon ;

reload drive letters when you refresh

you could insert or remove a disc and it'd vanish.